backend/requirements.txt
fastapi
uvicorn
python-docx
# weasyprint  # Removed due to DLL issues
xhtml2pdf
jinja2
groq
python-dotenv
pydantic
